09: Update init to be robust against federated roles
awscred doesn't support federated roles, and when faced with such config
it will just hang causing the tests to timeout and error.  We make this
scenario obvious, but right now there's no fix and you have to
workaround the issue by creating a dedicated IAM user via the console
with the necessary privs.